fix: moved endorse button to generic components
