+++ src/features/account/components/index.ts
@@ -3,5 +3,4 @@ export * from './reviews/ReviewsTabPanel';
 export * from './trusted-circle/TrustedCircleTabPanel';
 export * from './snaps/MySnapsTabPanel';
 export * from './AccountInfo';
-export * from './TEEndorsementButton';
 export * from './modals';
